man: Remove AUTHOR sections
These are strongly discouraged by the man-pages project, because they
are redundant with the copyright notices, and tend to get out-of-sync.
In most cases removed here these where documenting the authorship of
the programs not manual pages themselves, but when they were these have
been moved into comments.

In addition their format is inconsistent. Just remove them in the same
way we did for man pages in the past.